fix/ redux 리팩토링 & 수정
