+++ client/src/components/hire/HireArticle.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { CardProps, ServerData } from "./CardProps";
 import styled from "styled-components";
 import { useSelector } from "react-redux";
-import { RootState } from "../../util/store";
+import { RootState } from "../../util/redux";
 import axios from "axios";
 
 const mapDataToCardProps = (data: ServerData): CardProps => {
@@ -54,12 +54,14 @@ const HireArticle: React.FC = () => {
   }, []);
 
   const selectedCategory = useSelector(
-    (state: RootState) => state.selectedCategory
+    (state: RootState) => state.DropDown.selectedCategory
   );
   const selectedLocation = useSelector(
-    (state: RootState) => state.selectedLocation
+    (state: RootState) => state.DropDown.selectedLocation
   );
-  const selectedTag = useSelector((state: RootState) => state.selectedTag);
+  const selectedTag = useSelector(
+    (state: RootState) => state.DropDown.selectedTag
+  );
   const filteredCards =
     selectedCategory === "카테고리" &&
     selectedLocation === "지역" &&
