+++ client/src/util/redux/DropDown.ts
@@ -1,0 +1,52 @@
+// 액션 객체 타입
+type DropDownActions =
+  | { type: "SELECT_CATEGORY"; payload: string }
+  | { type: "SELECT_LOCATION"; payload: string }
+  | { type: "SELECT_TAG"; payload: string };
+
+// 상태 타입
+type DropDownState = {
+  selectedCategory: string;
+  selectedLocation: string;
+  selectedTag: string;
+};
+
+// 초기값
+const initialState: DropDownState = {
+  selectedCategory: "카테고리",
+  selectedLocation: "지역",
+  selectedTag: "",
+};
+
+//액션 생성 함수 선언
+export const selectCategory = (category: string) => ({
+  type: "SELECT_CATEGORY",
+  payload: category,
+});
+
+export const selectLocation = (location: string) => ({
+  type: "SELECT_LOCATION",
+  payload: location,
+});
+
+export const selectTag = (tag: string) => ({
+  type: "SELECT_TAG",
+  payload: tag,
+});
+
+// 리듀서
+export const DropDown = (
+  state = initialState,
+  action: DropDownActions
+): DropDownState => {
+  switch (action.type) {
+    case "SELECT_CATEGORY":
+      return { ...state, selectedCategory: action.payload };
+    case "SELECT_LOCATION":
+      return { ...state, selectedLocation: action.payload };
+    case "SELECT_TAG":
+      return { ...state, selectedTag: action.payload };
+    default:
+      return state;
+  }
+};
